feat: log out func
