+++ components/LogoutButton.tsx
@@ -1,16 +1,7 @@
-"use client";
-import { createClient } from "@/utils/supabase/server";
-import { redirect } from "next/navigation";
+import { signOut } from "@/app/action";
 import { Button } from "@/components/ui/button";
 
 export default function LogoutButton() {
-  const signOut = async () => {
-    "use server";
-
-    const supabase = createClient();
-    await supabase.auth.signOut();
-    return redirect("/");
-  };
   return (
     <Button variant="destructive" onClick={signOut}>
       Logout
